add(assignees.html): add error handling for assignee deletion

+++ week-08/day-2/todoexternaldb/src/main/java/com/greenfoxacademy/todoexternaldb/controllers/AssigneeController.java
@@ -1,7 +1,7 @@
 package com.greenfoxacademy.todoexternaldb.controllers;
 
 import com.greenfoxacademy.todoexternaldb.model.Assignee;
-import com.greenfoxacademy.todoexternaldb.service.AssigneeService;
+import com.greenfoxacademy.todoexternaldb.service.IAssigneeService;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -10,18 +10,16 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
-import org.springframework.web.servlet.ModelAndView;
 import org.springframework.web.servlet.mvc.support.RedirectAttributes;
-import org.springframework.web.servlet.view.RedirectView;
 
 
 @Controller
 @RequestMapping(value = {"/", "/todo"})
 public class AssigneeController {
 
-  private AssigneeService assigneeService;
+  private IAssigneeService assigneeService;
 
-  AssigneeController(AssigneeService assigneeService) {
+  AssigneeController(IAssigneeService assigneeService) {
 
     this.assigneeService = assigneeService;
   }
@@ -38,10 +36,11 @@ public class AssigneeController {
   }
 
   @PostMapping("/assignees/add")
-  public String addAssignee(@RequestParam String name, String email, RedirectAttributes redirectAttributes, Model model) {
+  public String addAssignee(@RequestParam String name, String email,
+      RedirectAttributes redirectAttributes) {
     String assigneeError;
     if (assigneeService.findAssigneeByName(name) != null) {
-      assigneeError = "true";
+      assigneeError = "It looks like this assignee already exists. Try another one!";
     } else {
       assigneeService.saveAssignee(new Assignee(name, email));
       assigneeError = "";
@@ -51,8 +50,15 @@ public class AssigneeController {
   }
 
   @GetMapping(value = "assignees/{id}/delete")
-  public String deleteTodo(@PathVariable Long id) {
-    assigneeService.deleteAssignee(id);
+  public String deleteTodo(@PathVariable Long id, RedirectAttributes redirectAttributes) {
+    String deletionError;
+    if (assigneeService.getAssignee(id).getTodoList().size() > 0) {
+      deletionError = "It looks like this assignee still has some ToDos.";
+    } else {
+      deletionError = "";
+      assigneeService.deleteAssignee(id);
+    }
+    redirectAttributes.addFlashAttribute("assigneeError", deletionError);
     return "redirect:/todo/assignees";
   }
